% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_syst.R
\name{get_syst}
\alias{get_syst}
\title{Determine the appropriate GUI to instantiate}
\usage{
get_syst(rstudio = TRUE)
}
\arguments{
\item{rstudio}{Logical. Should RStudio dialog automatically be used if
available?}
}
\value{
A character scalar giving either "RStudio" or the result of
\code{Sys.info()["sysname"]}
}
\description{
Determine the appropriate GUI to instantiate
}
\keyword{internal}
