\name{guiDlgColor}
\alias{guiDlgColor}

\title{ A color selection dialog box }
\description{
  Allows to select a color graphically
}
\usage{
guiDlgColor(...)
}

\arguments{
  \item{\dots}{ further arguments to the color dialog box }
}

\value{
  Returns selected colors
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgFont}}, \code{\link{guiDlgGraphOptions}} }

\examples{
#### TO DO...
}

\keyword{ misc }
