\name{guiDlgFont}
\alias{guiDlgFont}

\title{ A font selection dialog box }
\description{
  Select a font and some attributes.
}

\usage{
guiDlgFont(...)
}

\arguments{
  \item{\dots}{ further arguments to the function. }
}

\value{
  A string describing the selected font.
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgOptions}} }

\examples{
### TODO...
}

\keyword{ misc }

\concept{ GUI API dialog boxes }
