\name{guiDlgFormula}
\alias{guiDlgFormula}

\title{ A formula dialog box }
\description{
  Eases the creation of formulas
}
\usage{
guiDlgFormula(...)
}

\arguments{
  \item{\dots}{ further arguments to the function }
}

\value{
  A formula build with this dialog box
}
\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgFunction}} }

\examples{
#### TO DO...
}

\keyword{ misc }
