\name{guiDlgFunction}
\alias{guiDlgFunction}

\title{ A dialog box prompting for function arguments }
\description{
  Given a function, this dialog box prompts for values for its arguments
}

\usage{
guiDlgFunction(fun, template = NULL, maxargs = 7, var = "res", width = 40,
    labelwidth = 10, displayit = TRUE, execfun = getOption("guiExecFun"))
}

\arguments{
  \item{fun}{ The name of the function to prompt for }
  \item{template}{ An optional template for the dialog box }
  \item{maxargs}{ The maximum number of arguments to include in the dialog box }
  \item{var}{ The default name of the variable where to assign result }
  \item{width}{ The width of the dialog box }
  \item{labelwidth}{ The width reserved for labels }
  \item{displayit}{ Do we display it, or just construct it? }
  \item{execfun}{ The function to use to execute resulting code }
}

\value{
  A line of R code generated by the dialog box
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgInput}}, \code{\link{guiDlgVarSel}} }

\examples{
#### TO DO...
}

\keyword{ misc }
