\name{guiDlgGraphOptions}
\alias{guiDlgGraphOptions}

\title{ A graph options dialog box }
\description{
  Change various graph options at once with this dialog box.
}
\usage{
guiDlgGraphOptions(\dots)
}

\arguments{
  \item{\dots}{ further arguments to the function. }
}

\value{
  This function is invoked for its side effect to change graph options with a
  GUI dialog box.
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgOptions}} }

\examples{
### TODO...
}

\keyword{ misc }

\concept{ GUI API dialog boxes }
