\name{guiDlgOptions}
\alias{guiDlgOptions}

\title{ A general options dialog box }
\description{
  Display and allow to change various options.
}
\usage{
guiDlgOptions(\dots)
}

\arguments{
  \item{\dots}{ further arguments to the function. }
}

\value{
  This function is invoked for its side effect to change options through
  a dialog box.
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgGraphOptions}} }

\examples{
### TODO...
}

\keyword{ misc }

\concept{ GUI API dialog boxes }
