\name{svDialogs-package}
\alias{svDialogs-package}
\alias{svDialogs}
\docType{package}

\title{ SciViews GUI API - Dialog boxes }
\description{
  Rapidly construct dialog boxes for your GUI, including an automatic
  function assistant.
}
\details{
  \tabular{ll}{
    Package: \tab svDialogs\cr
    Type: \tab Package\cr
    Version: \tab 0.9-43\cr
    Date: \tab 2010-09-25\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API dialog boxes }
