\name{display}
\alias{display}

\title{ The display method (for flexible dialog boxes) }
\description{
  Display a dialog box previously calculated. For instance, display a flexible
  dialog box.
}

\usage{
display(x, \dots)
}

\arguments{
  \item{x}{ an object that accepts a display method. }
  \item{\dots}{ further arguments passed to the display method. }
}

\value{
  The R code constructed by this dialog box.
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlg}}, \code{\link{guiEval}} }

\examples{
### TODO...
}

\keyword{misc}

\concept{ GUI API dialog boxes }
