\name{guiDlgAssistant}
\alias{guiDlgAssistant}

\title{ An assistant dialog box }
\description{
  The assistant allows step by step execution of a given task
}
\usage{
guiDlgAssistant(...)
}

\arguments{
  \item{\dots}{ arguments for the assistant dialog box. }
}

\value{
  This function is invoked for its side effect: to display an assistant dialog
  box.
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgFunction}} }

\examples{
### TODO...
}

\keyword{ misc }

\concept{ GUI API dialog boxes }
