\name{guiDlgFunction}
\alias{guiDlgFunction}

\title{ A dialog box prompting for function arguments }
\description{
  Given a function, this dialog box prompts for values for its arguments.
}

\usage{
guiDlgFunction(fun, template = NULL, maxargs = 7, var = "res", width = 40,
    labelwidth = 10, displayit = TRUE, execfun = getOption("guiExecFun"))
}

\arguments{
  \item{fun}{ the name of the function to prompt for. }
  \item{template}{ an optional template for the dialog box. }
  \item{maxargs}{ the maximum number of arguments to include in the dialog box. }
  \item{var}{ the default name of the variable where to assign result. }
  \item{width}{ the width of the dialog box. }
  \item{labelwidth}{ the width reserved for labels. }
  \item{displayit}{ do we display it, or just construct it? }
  \item{execfun}{ the function to use to execute resulting code. }
}

\value{
  A line of R code generated by the dialog box.
}

\author{Philippe Grosjean (\email{phgrosjean@sciviews.org})}

\seealso{ \code{\link{guiDlgInput}}, \code{\link{guiDlgVarSel}} }

\examples{
### TODO...
}

\keyword{ misc }

\concept{ GUI API dialog boxes }
