% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svDialogs-package.R
\docType{package}
\name{svDialogs-package}
\alias{svDialogs-package}
\title{SciViews - Standard Dialog Boxes for R}
\description{
Rapidly construct standard dialog boxes for your GUI, including
message boxes, input boxes, list, file or dir selection, ... In case R cannot
display GUI dialog boxes, a simpler command line version of these interactive
elements is also provided as fallback solution.
}
\section{Important functions}{

\itemize{
\item \code{\link[=dlg_dir]{dlg_dir()}} select a directory,
}
}

